split nft-utils and move module specific code into module and remaining into demo-rollup
