+++ Cargo.toml
@@ -19,9 +19,7 @@ members = [
 
     "utils/zk-cycle-macros",
     "utils/zk-cycle-utils",
-    "utils/nft-utils",
     "utils/bashtestmd",
-    "utils/nft-utils",
 
     "module-system/sov-cli",
     "module-system/sov-modules-stf-template",
