+++ examples/demo-rollup/Cargo.toml
@@ -19,11 +19,12 @@ sov-stf-runner = { path = "../../full-node/sov-stf-runner" }
 sov-rollup-interface = { path = "../../rollup-interface" }
 sov-modules-stf-template = { path = "../../module-system/sov-modules-stf-template" }
 sov-modules-api = { path = "../../module-system/sov-modules-api" }
+sov-nft-module = { path = "../../module-system/module-implementations/sov-nft-module" }
 demo-stf = { path = "../demo-stf" }
 methods = { path = "../demo-prover/methods" }
+borsh = { workspace = true, features = ["bytes"]}
 async-trait = { workspace = true, optional = true }
 anyhow = { workspace = true, optional = true }
-borsh = { workspace = true, features = ["bytes"], optional = true }
 jsonrpsee = { workspace = true, features = ["http-client", "server"], optional = true }
 serde = { workspace = true, features = ["derive"], optional = true }
 serde_json = { workspace = true, optional = true }
@@ -78,7 +79,7 @@ native = ["anyhow", "jsonrpsee", "serde", "serde_json", "tracing", "tokio", "tra
 	"demo-stf/native", "sov-modules-stf-template/native", "sov-risc0-adapter/native", "sov-modules-api/native",
 	"sov-state/native", "sov-cli", "clap", "sov-celestia-adapter/native", "sov-db", "sov-sequencer", "sov-stf-runner/native",
 	"sov-modules-api/native", "sov-rollup-interface/native"]
-bench = ["native", "async-trait", "borsh", "hex"]
+bench = ["native", "async-trait", "hex"]
 offchain = ["demo-stf/offchain"]
 
 [[bench]]
@@ -98,6 +99,11 @@ required-features = ["bench"]
 [[bin]]
 name = "sov-cli"
 path = "src/sov-cli/main.rs"
+required-features = ["native"]
+
+[[bin]]
+name = "sov-nft-script"
+path = "src/bin/sov_nft_script.rs"
 required-features = ["native"]
 
 [[bin]]
